Language Training and Refugees' Integration
We evaluate a Danish reform focused on improving language training for those granted refugee status on or after January 1, 1999. Using a Regression Discontinuity Design we find a significant, permanent, positive effect on earnings. This effect emerged after completion of language classes and was accompanied by additional schooling and higher probability of working in complex jobs, consistent with language training, rather than other minor aspects of the reform, producing the results.